pyproject.toml
[project]
name = "hellopy-backend"
version = "0.1.0"
description = "hellopy-backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "django>=5.1.5",
   "django-jazzmin>=3.0.1",
   "djangorestframework>=3.15.2",
   "rich>=13.9.4",
   "pillow>=11.1.0",
   "django-ckeditor-5",
   "django-filter>=25.1",
]

[tool.ruff]
target-version = "py312"
extend-exclude = [
   "**/*.yaml",
   "**/.gitignore",
   "**/.ruff_cache/*",
   "**/.mypy_cache.*",
   "**/__pycache__/*",
   "**/assets/*",
   "**/.vscode/*",
   "**/.git/*",
   "**/.idea/*",
   "**/temp/*",
   "**/*venv*/*",
]
line-length = 100
fix = true
indent-width = 4

[tool.ruff.lint]
select = ["I", "E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[dependency-groups]
dev = ["drf-spectacular>=0.28.0", "pre-commit>=4.1.0", "ruff>=0.9.4"]
